Giftcards: Validate amounts
Make use of the new HpsInputValidation.check_amount method. Fix the  argument error bug in the method - the exception requires at least a message to be provided. Pass client transaction id to the doTransaction method now that it excepts that argument. GiftCardService methods requiring a gift card and an amount now ask for these 2 arguments in the same position; this also allows for currency default to work properly; update tests to reflect this change. With previous change, we can now test exceptions from those methods and expect that they throw when the amount is less than zero.